Removed the constructor taking monomer groups as arguments
Instead, the analyzer uses the whole system, and places it
into a monomer group in initialize(), if the groups have not
been set explicitly using the setter function provided.
This has the advantage that one can make an own Rg analyzer class by
inheriting from AnalyzerRadiusOfGyration, and explicitly set
the groups there in the initialize. Otherwise one can use the same
execute and cleanup as provided here.